requirements.txt
Flask
flask-cors
scikit-learn
pandas
joblib
mysql-connector-python
